Add support for branch previews of Jekyll site in GitHub Actions and Pages
Introduce a new GitHub Actions workflow that reacts to pushes to branches.
When changes are pushed to a branch we build the Jekyll site and deploy it
in a subdirectory of GitHub Pages.

There are 3 key changes here:

1. To support multiple versions of the same site we need to deploy it to a
subsirectory.
2. To support Jekyll running in a subdirectory we need to set the Jekyll base
url explicitly in Jekyll _config.yml.
3. Since we build the Jekyll site ourselves we do not want GitHub Pages to do it

Use a set of GitHub Actions to achieve the process that are also used other
places in the DPL NEXT project (peaceiris/actions-gh-pages) or provided by
GitHub (actions/configure-pages, actions/jekyll-build-pages).