Add support for deployments of branch previews
This creates nice links to the deployed versions of the documentation from
GitHub repositories and pull requests.

When a branch is deployed then create a corresponding GitHub environment
(this happens implictly) and deploy to it. When deployment has finished then
update the deployment status accordingly.

Use bobheadxi/deployments which is also used elsewhere to manage this.