Wait for github pages preview to become available before deployment as completed
The deployment we initiate finishes sooner than new branch previews are
actually available in GitHub Actions as GitHubs own deployment needs to finish.

Wait for the process to finish before we mark the deployment as active.

We use a separate action here than elsewhere. nev7n/wait-for_response takes
about 1 minute to build for each run. This is too slow. The one used here
relies on a popular NPM package and should be much quicker.